+++ wazo_calld/plugins/meetings/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2021 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2021-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_confd_client import Client as ConfdClient
@@ -7,6 +7,7 @@ from wazo_amid_client import Client as AmidClient
 from .http import (
     MeetingParticipantsResource,
     MeetingParticipantsUserResource,
+    MeetingStatusGuestResource,
 )
 from .bus_consume import MeetingsBusEventHandler
 from .notifier import MeetingsNotifier
@@ -28,7 +29,7 @@ class Plugin:
         token_changed_subscribe(amid_client.set_token)
         token_changed_subscribe(confd_client.set_token)
 
-        meetings_service = MeetingsService(amid_client, ari.client, confd_client)
+        meetings_service = MeetingsService(amid_client, ari.client, confd_client, config)
         notifier = MeetingsNotifier(bus_publisher)
         bus_event_handler = MeetingsBusEventHandler(
             confd_client, notifier, meetings_service
@@ -43,5 +44,10 @@ class Plugin:
         api.add_resource(
             MeetingParticipantsUserResource,
             '/users/me/meetings/<meeting_uuid>/participants',
+            resource_class_args=[meetings_service],
+        )
+        api.add_resource(
+            MeetingStatusGuestResource,
+            '/guests/me/meetings/<meeting_uuid>/status',
             resource_class_args=[meetings_service],
         )